fix(rewards): unify weekly behavior board with rewards page
Extract BehaviorTracking into a shared component and compose the
rewards page as a unified task-to-reward experience: weekly star grid
on top, unlockable rewards below. The child sees the direct link
between earning stars and unlocking rewards on the same page.

- Extract BehaviorTracking + BehaviorForm into components/barkley/
- Rewards page now embeds the full weekly behavior tracking grid
- Add visual connector showing cumulative star count between zones
- Fix cache invalidation: toggle star now refreshes star count
- Simplify /barkley to Programme-only page
- Rename nav label to "Programme Barkley"
